reemplazo de clases por su implementacion en herramientas_desarrollo. #13.

+++ modelo/include/Fecha.h
@@ -53,7 +53,7 @@ public:
 
     virtual std::string prefijoGrupo();
 
-    virtual unsigned int hashcode();
+    virtual unsigned long long int hashcode();
 
     // metodos de IEntidad
