+++ modelo/include/RelacionesMedio.h
@@ -40,7 +40,7 @@ public:
 
     virtual std::string prefijoGrupo();
 
-    virtual unsigned int hashcode();
+    virtual unsigned long long int hashcode();
 
     // metodos de IContieneJson
