+++ modelo/include/RelacionesSeccion.h
@@ -36,7 +36,7 @@ public:
 
     virtual std::string prefijoGrupo();
 
-    virtual unsigned int hashcode();
+    virtual unsigned long long int hashcode();
 
     // metodos de IContieneJson
