+++ modelo/include/Termino.h
@@ -50,7 +50,7 @@ public:
 
 	virtual std::string prefijoGrupo();
 
-	virtual unsigned int hashcode();
+	virtual unsigned long long int hashcode();
 
     // metodos de IEntidad
